+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Fixed issue with article scroll progress tracking if `article.elementFn` callback wasn't set or didn't return any element. 
 - Upgraded to Laravel 6. remp/remp#491
 - Added UI for `TagCategory`. remp/remp#898
+- Fixed filter bugs in article datatable. remp/remp#921
 
 ### [Campaign]